chore(deps): Bump k8s dependencies to v1.26.11 (#3211)
* Bump k8s dependencies to v1.26.0.

This addresses a few CVEs, but was slightly challenging because of some Kubernetes API changes.

In particular, the Ingress type signatures changed. This required some surgery on a function that
wrapped the extensionsv1beta1 Ingress and v1 Ingress types, attempting to provide some compatibility.

The API changes were kind of tough to deal with, but fortunately the only field required by anything else
in this project was the Hostname field, so I changed the function to just normalize the hostnames into
a slice.

Signed-off-by: Dan Lorenc <dlorenc@chainguard.dev>

* Update codegen.

Signed-off-by: Dan Lorenc <dlorenc@chainguard.dev>

* bump the replace deps, and re-run codegen

Signed-off-by: Zach Aller <zachaller@users.noreply.github.com>

* bump to 0.26.11 and possible fix for new crd properties

Signed-off-by: Zach Aller <zachaller@users.noreply.github.com>

* remove local changes

Signed-off-by: Zach Aller <zachaller@users.noreply.github.com>

* switch alb test off of extensions and on to v1

Signed-off-by: Zach Aller <zachaller@users.noreply.github.com>

---------

Signed-off-by: Dan Lorenc <dlorenc@chainguard.dev>
Signed-off-by: Zach Aller <zachaller@users.noreply.github.com>
Co-authored-by: Zach Aller <zachaller@users.noreply.github.com>
Signed-off-by: ashutosh16 <11219262+ashutosh16@users.noreply.github.com>